refactor: AddBtn을 AddStoreBtn으로 수정
- StoreAddBtnController 의 이벤트 리스너를 AddStoreBtn으로 이동, 해당 모듈 삭제
