+++ src/components/AddStoreBtn/style.css
@@ -1,4 +1,4 @@
-add-btn button {
+add-store-btn button {
   cursor: pointer;
   box-sizing: border-box;
   border: none;
